requirements.txt
torch
torchvision
pillow
numpy
gradio